# services/streaming-service/openapi.yaml
openapi: 3.0.3

info:
  title: Streaming/Playback Service API
  description: >
    Provides media-location lookup and playback-position tracking.
    Media files are pre-encoded MP4 assets stored in Azurite during local
    development. The service does not perform transcoding, adaptive bitrate
    streaming, HLS/DASH manifest generation, DRM, or media segmentation.

    The media URL returned by this API points to the stored media object.
    Clients may use HTTP Range requests against that URL for partial media
    retrieval.

    SQL Server (StreamingDb) stores media metadata and playback positions.
    See ADR 0006 (docs/adr/0006-streaming-data-store.md) for that decision
    and its explicitly unresolved open questions (retention policy,
    cross-service titleId validation against Catalog).

    Known contract-level assumptions — not yet formally decided:
      - `titleId` existence is checked only against this service's own
        registered media records in StreamingDb. This service does not
        call Catalog to verify the title exists there. A 404 on
        GET .../media means "no media asset registered here," not
        "title does not exist in Catalog."
      - Playback positions are retained indefinitely. No TTL or scheduled
        cleanup job is implemented.
  version: "1.1"

servers:
  - url: /playback
    description: >
      Direct service access in Phases 1–3; reached through the API Gateway
      from Phase 4 onward.

tags:
  - name: Playback
  - name: Health

paths:

  /api/v1/playback/{titleId}/media:
    get:
      tags:
        - Playback
      summary: Get media information for a title
      description: >
        Looks up the stored media object for a title from StreamingDb.
        Existence is checked against this service's own records only —
        not against Catalog. See info.description for the open
        cross-service validation question.
      operationId: getMedia
      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/TitleId'

      responses:

        '200':
          description: Media information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaResponse'

        '404':
          description: >
            No media asset is registered in StreamingDb for this titleId.
            Does not imply the title is absent from Catalog.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'


  /api/v1/playback/{titleId}/position:
    get:
      tags:
        - Playback

      summary: Get saved playback position

      operationId: getPosition

      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/TitleId'

        - name: profileId
          in: query
          required: true
          schema:
            type: string
            format: uuid

      responses:

        '200':
          description: Saved playback position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackPositionResponse'

        '404':
          description: No saved position exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'


    put:
      tags:
        - Playback

      summary: Save playback position
      description: >
        Upserts the playback position for the given title and profile.
        Positions are retained indefinitely — no retention/expiry policy
        is implemented (open item, see ADR 0006).

      operationId: savePosition

      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/TitleId'

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavePositionRequest'

      responses:

        '204':
          description: Playback position saved

        '400':
          $ref: '#/components/responses/ValidationError'


  /health:
    get:
      tags:
        - Health

      summary: Health check

      operationId: getHealth

      responses:
        '200':
          description: Service is healthy


components:

  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  parameters:

    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        Correlation identifier propagated across service calls.
        Generated by the service if omitted.
      schema:
        type: string
        format: uuid

    TitleId:
      name: titleId
      in: path
      required: true
      description: >
        Identifier shared with Catalog's title id. This service does not
        verify the id against Catalog — see info.description.
      schema:
        type: string


  responses:

    ValidationError:
      description: One or more fields failed validation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'


  schemas:

    MediaResponse:
      type: object

      required:
        - titleId
        - mediaUrl
        - contentType
        - durationSeconds

      properties:

        titleId:
          type: string

        mediaUrl:
          type: string
          format: uri
          description: >
            URL of the pre-encoded media object stored in Azurite during
            local development.

        contentType:
          type: string
          example: video/mp4

        durationSeconds:
          type: integer
          minimum: 0


    PlaybackPositionResponse:
      type: object

      required:
        - titleId
        - profileId
        - positionSeconds
        - updatedAtUtc

      properties:

        titleId:
          type: string

        profileId:
          type: string
          format: uuid

        positionSeconds:
          type: integer
          minimum: 0

        updatedAtUtc:
          type: string
          format: date-time


    SavePositionRequest:
      type: object

      required:
        - profileId
        - positionSeconds

      properties:

        profileId:
          type: string
          format: uuid

        positionSeconds:
          type: integer
          minimum: 0


    ProblemDetails:
      type: object

      properties:

        type:
          type: string

        title:
          type: string

        status:
          type: integer

        detail:
          type: string

        instance:
          type: string

        correlationId:
          type: string
          format: uuid


    ValidationProblemDetails:
      allOf:

        - $ref: '#/components/schemas/ProblemDetails'

        - type: object
          properties:

            errors:
              type: object

              additionalProperties:
                type: array

                items:
                  type: string